requirements.txt
numpy~=2.3.1
onnxruntime~=1.22.0
pillow~=11.2.1
fastapi~=0.115.13
uvicorn
rsqlalchemy
python-dotenv

SQLAlchemy~=2.0.41